# TrafficPatrolProject-bugfix/002implementations/api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Trafficpatrol API",
        "description": "Trafficpatrol API documentation",
        "termsOfService": "",
        "contact": {
            "name": "",
            "url": "",
            "email": ""
        },
        "license": {
            "name": "unlicensed",
            "url": ""
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/contents/{id}": {
            "get": {
                "tags": [
                    "content-resource"
                ],
                "operationId": "getContent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContentDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "content-resource"
                ],
                "operationId": "updateContent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContentDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContentDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "content-resource"
                ],
                "operationId": "deleteContent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "tags": [
                    "content-resource"
                ],
                "operationId": "partialUpdateContent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContentDTO"
                            }
                        },
                        "application/merge-patch+json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContentDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContentDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/content-blobs/{id}": {
            "get": {
                "tags": [
                    "content-blob-resource"
                ],
                "operationId": "getContentBlob",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContentBlobDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "content-blob-resource"
                ],
                "operationId": "updateContentBlob",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContentBlobDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContentBlobDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "content-blob-resource"
                ],
                "operationId": "deleteContentBlob",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "tags": [
                    "content-blob-resource"
                ],
                "operationId": "partialUpdateContentBlob",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContentBlobDTO"
                            }
                        },
                        "application/merge-patch+json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContentBlobDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContentBlobDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/contents": {
            "get": {
                "tags": [
                    "content-resource"
                ],
                "operationId": "getAllContents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ContentDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "content-resource"
                ],
                "operationId": "createContent",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContentDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContentDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/content-blobs": {
            "get": {
                "tags": [
                    "content-blob-resource"
                ],
                "operationId": "getAllContentBlobs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ContentBlobDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "content-blob-resource"
                ],
                "operationId": "createContentBlob",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContentBlobDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContentBlobDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AppUserDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "phoneNumber": {
                        "maximum": 1000000000,
                        "minimum": 1000000000,
                        "type": "integer",
                        "format": "int64"
                    },
                    "address": {
                        "type": "string"
                    },
                    "pinCode": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserDTO"
                    }
                }
            },
            "ContentDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "contentType": {
                        "type": "string"
                    },
                    "uploadDateTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "type": "string"
                    },
                    "appUser": {
                        "$ref": "#/components/schemas/AppUserDTO"
                    }
                }
            },
...
            "UserDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "login": {
                        "type": "string"
                    }
                }
            },
...
            "ContentBlobDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "contentBlob": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "byte"
                        }
                    },
                    "contentBlobContentType": {
                        "type": "string"
                    },
                    "content": {
                        "$ref": "#/components/schemas/ContentDTO"
                    }
                }
            },
...
        }
    }
}